html/semantics/embedded-content-0/the-video-element/video_defaultMuted_reflect.html

<!DOCTYPE html>
<html>
  <head>
    <title>Video Test: video_defaultMuted_check</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://dev.w3.org/html5/spec/single-page.html#dom-media-defaultmuted" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check that the defaultMuted reflects the muted content attribute in the video element but has no dynamic effect" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <video id="m" muted>The user agent doesn't support media element.</video>
    <video id="v">The user agent doesn't support media element.</video>
    <script type="text/javascript">
        var media = document.getElementById("m");
        var video = document.getElementById("v");
        
        test(function() {
            assert_true(media.defaultMuted, "media.defaultMuted of value");
        }, "Check if video.defaultMuted returns true when the muted attribute is present in the video element");
        
        test(function() {
            media.muted = false;
            assert_true(media.defaultMuted, "media.defaultMuted of value");
        }, "Check if video.defaultMuted returns true after the muted attribute is set to false when the muted attribute is present in the video element");
        
        test(function() {
            assert_false(video.defaultMuted, "media.defaultMuted of value");
        }, "Check if video.defaultMuted returns false when the muted attribute is absent in the video element");
        
        test(function() {
            video.muted = true;
            assert_false(video.defaultMuted, "media.defaultMuted of value");
        }, "Check if video.defaultMuted returns false after the muted attribute is set to true when the muted attribute is absent in the video element");
    </script>
  </body>
</html>
